+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # 0.0.1 [2023-04-28]
 
-- Initial release of the post_to_facebook package.
-- The package allows posting messages on Facebook on behalf of a user or page using the Facebook Graph API.
+* Initial release of the post_to_facebook package.
+* The package allows posting messages on Facebook on behalf of a user or page using the Facebook Graph API.
+
+# 0.0.2 [2023-04-28]
+
+* Able to post images using image url